Make similar fixes for snapshots
