+++ src/main/java/bio/terra/service/job/JobService.java
@@ -201,8 +201,6 @@ public class JobService {
 
         } catch (StairwayException stairwayEx) {
             throw new InternalStairwayException("Stairway initialization failed", stairwayEx);
-        } catch (InterruptedException ex) {
-            throw new JobServiceShutdownException("Job service interrupted", ex);
         }
     }
