+++ src/main/java/bio/terra/service/snapshot/flight/create/SnapshotAuthzTabularAclStep.java
@@ -1,9 +1,9 @@
-package bio.terra.service.dataset.flight.create;
+package bio.terra.service.snapshot.flight.create;
 
 import bio.terra.common.exception.PdaoException;
-import bio.terra.service.dataset.Dataset;
-import bio.terra.service.dataset.DatasetService;
-import bio.terra.service.dataset.flight.DatasetWorkingMapKeys;
+import bio.terra.service.snapshot.Snapshot;
+import bio.terra.service.snapshot.SnapshotService;
+import bio.terra.service.snapshot.flight.SnapshotWorkingMapKeys;
 import bio.terra.service.tabulardata.google.BigQueryPdao;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.FlightMap;
@@ -16,34 +16,35 @@ import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.List;
 import java.util.UUID;
 
-public class CreateDatasetAuthzGetCurrentPrimaryAclStep implements Step {
-    private static final Logger logger = LoggerFactory.getLogger(CreateDatasetAuthzGetCurrentPrimaryAclStep.class);
-
+public class SnapshotAuthzTabularAclStep implements Step {
     private final BigQueryPdao bigQueryPdao;
-    private final DatasetService datasetService;
+    private final SnapshotService snapshotService;
+    private static final Logger logger = LoggerFactory.getLogger(SnapshotAuthzTabularAclStep.class);
 
-    public CreateDatasetAuthzGetCurrentPrimaryAclStep(
-        BigQueryPdao bigQueryPdao,
-        DatasetService datasetService) {
+    public SnapshotAuthzTabularAclStep(BigQueryPdao bigQueryPdao,
+                                       SnapshotService snapshotService) {
         this.bigQueryPdao = bigQueryPdao;
-        this.datasetService = datasetService;
+        this.snapshotService = snapshotService;
     }
 
     @Override
     public StepResult doStep(FlightContext context) throws InterruptedException {
         FlightMap workingMap = context.getWorkingMap();
-        UUID datasetId = workingMap.get(DatasetWorkingMapKeys.DATASET_ID, UUID.class);
-        List<String> policyEmails = workingMap.get(DatasetWorkingMapKeys.POLICY_EMAILS, List.class);
-        Dataset dataset = datasetService.retrieve(datasetId);
+        UUID snapshotId = workingMap.get(SnapshotWorkingMapKeys.SNAPSHOT_ID, UUID.class);
+        Snapshot snapshot = snapshotService.retrieve(snapshotId);
+
+        String readersPolicyEmail = workingMap.get(SnapshotWorkingMapKeys.POLICY_EMAIL, String.class);
+
         try {
-            bigQueryPdao.grantReadAccessToDataset(dataset, policyEmails);
+            bigQueryPdao.addReaderGroupToSnapshot(snapshot, readersPolicyEmail);
         } catch (BigQueryException ex) {
             // TODO: OK, so how do I know it is an IAM error and not some other error?
             //  First, I just scan for IAM in the message text and I log the BigQueryError reason. Then I
             //  can come back to this code and be more explicit about the reason.
+            //  This is a cut'n'paste from CreateDatasetAuthzPrimaryDataStep. When we fix it there,
+            //  we should also fix it here.
             BigQueryError bqError = ex.getError();
             if (bqError != null) {
                 logger.info("BigQueryError: reason=" + bqError.getReason() + " message=" + bqError.getMessage());
@@ -51,16 +52,15 @@ public class CreateDatasetAuthzGetCurrentPrimaryAclStep implements Step {
             if (StringUtils.contains(ex.getMessage(), "IAM")) {
                 return new StepResult(StepStatus.STEP_RESULT_FAILURE_RETRY, ex);
             }
-            throw new PdaoException("Caught BQ exception while granting read access to dataset", ex);
+            throw new PdaoException("Caught BQ exception while granting read access to snapshot", ex);
         }
         return StepResult.getStepResultSuccess();
     }
 
     @Override
-    public StepResult undoStep(FlightContext context) throws InterruptedException {
-        FlightMap workingMap = context.getWorkingMap();
-        UUID datasetId = workingMap.get(DatasetWorkingMapKeys.DATASET_ID, UUID.class);
-
+    public StepResult undoStep(FlightContext context) {
+        // We do not try to undo the ACL set, because we expect the entire BQ dataset to get deleted,
+        // taking the ACLs with it if we are on this undo path.
         return StepResult.getStepResultSuccess();
     }
 }
